+++ src/components/manage/Blocks/LeadImage/LeadImageSidebar.test.jsx
@@ -32,6 +32,7 @@ test('renders a Lead Image block Sidebar component', () => {
         pathname="/news"
         onChangeBlock={() => {}}
         openObjectBrowser={() => {}}
+        onChangeField={() => {}}
       />
     </Provider>,
   );