+++ src/components/manage/Blocks/Listing/ListingBody.test.jsx
@@ -56,6 +56,7 @@ test('renders a ListingBody component', () => {
           properties={{ is_folderish: true }}
           block="1234"
           onChangeBlock={() => {}}
+          path={() => {}}
         />
       </MemoryRouter>
     </Provider>,